refactor(StreetMetaWidth): split up into separate view components
