+++ assets/scripts/streets/StreetMeta.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import StreetMetaWidth from './StreetMetaWidth'
+import StreetMetaWidthContainer from './StreetMetaWidthContainer'
 import StreetMetaAuthor from './StreetMetaAuthor'
 import StreetMetaDate from './StreetMetaDate'
 import StreetMetaGeotag from './StreetMetaGeotag'
@@ -7,7 +7,7 @@ import './StreetMeta.scss'
 
 const StreetMeta = (props) => (
   <div className="street-metadata">
-    <StreetMetaWidth />
+    <StreetMetaWidthContainer />
     <StreetMetaGeotag />
     <StreetMetaAuthor />
     <StreetMetaDate />
