+++ assets/scripts/streets/StreetMeta.scss
@@ -1,8 +1,7 @@
 @import '../../styles/variables.scss';
 
 .street-metadata {
-  padding-top: 7px;
-  padding-bottom: 40px;
+  padding-top: 0.5em;
   color: black;
   font-size: 13px;
   user-select: none;
@@ -12,41 +11,6 @@
   a {
     color: inherit;
   }
-}
-
-.street-width {
-  padding: 5px 0;
-}
-
-.street-width.street-width-editable {
-  cursor: pointer;
-
-  &:hover,
-  &:hover * {
-    color: darken($ui-colour, 40%);
-  }
-
-  &::after {
-    font-family: 'Verdana', sans-serif;
-    content: ' ▼';
-    padding-left: 5px;
-    font-size: 9px;
-    display: inline-block;
-    transform: scaleY(0.75);
-  }
-}
-
-.street-width-over {
-  color: $warning-colour;
-}
-
-.street-width-under {
-  color: rgb(96, 96, 96);
-  transition: color $environment-transition;
-}
-
-.street-width-select {
-  width: 150px;
 }
 
 body:not(.read-only) .street-metadata-author a:hover {
@@ -74,10 +38,6 @@ body.read-only .street-metadata-author a {
 body.dark-environs-invert-ui {
   .street-metadata {
     color: #fff;
-  }
-
-  .street-width-under {
-    color: #a0a0a0;
   }
 
   .street-metadata-author:not(:empty)::before,
