+++ assets/scripts/streets/StreetMetaWidthMenu.scss
@@ -1,0 +1,3 @@
+.street-width-select {
+  width: 150px;
+}
